refactor(messaging-adapters): clear the deferred review items
Reuse
- `buildExplorerDeepLink` moves to `base/message-formatter.ts`. It was
  byte-identical in both builders and this branch had already edited both
  copies once; the two deep-link test suites now exercise one function.
- `isBotAuthored` no longer sniffs Slack Events API fields (`bot_id`,
  `bot_profile`, the `bot_message` subtype) from the platform-neutral base.
  The base answers `false` and Slack states its own convention, as the
  JSDoc there always said platforms should.

Correctness
- Only `interactive` is subscribed in Socket Mode. The SDK emits the inner
  event type for an `events_api` envelope and the envelope type otherwise,
  and Slack labels every block action and view submission `interactive` — so
  the `block_actions`/`view_submission` subscriptions never fired. Had they,
  a view submission would have run the agent twice. Verified against
  @slack/socket-mode 2.0.6.
- One rejected file upload no longer aborts the remaining files in the reply.
- Agent-authored labels are escaped before going into Slack mrkdwn, where a
  `>` closed the `<url|label>` link early and leaked the rest as literal text.
- Truncation slices by code point, so a cut no longer strands half a surrogate
  pair.
- `versionNumber` resolves URL-first like `conversationId` and `artifactId`;
  a cold-load deep link could pair a URL artifact with the previous visit's
  version.

The test platform now declares its own bot marker rather than inheriting a
Slack-shaped default, which is the point of the layering change.
